fix(dashboard table view cell): fix odd animation on opening expanded state
Previously, adding or removing the expanded stack view to the cell's main stack view would not be animated, but the resizing of the cell to the expanded state would be. This meant, that at the start of the animation, the non expanded state would be initially shrunk to fix both itself and the expanded view within a non-expanded cell, or would expand to fix the size of the expanded cell, before the cell animated to the correct size. This would look odd, with the non-expanded content for the cell jumping around during the animation. This fix ensures that expanding the cell animates at the same rate as adding the expanded content.